% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pipini.R
\name{pipini}
\alias{pipini}
\title{Insert ->df%>%}
\usage{
pipini()
}
\value{
Insert ->df%>% written with the cursor moved to the line below
}
\description{
Insert ->df%>%
}
\examples{
Check and try the keyboard shortcut
}
