% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seletail.R
\name{seletail}
\alias{seletail}
\title{seletail function}
\usage{
seletail(df, n, selec = FALSE, exclu = FALSE)
}
\arguments{
\item{df}{A data frame.}

\item{n}{Number of variables to keep starting from the last variable.}

\item{selec}{A logical value indicating whether to print the names of the selected variables. Default is `FALSE`.}

\item{exclu}{A logical value indicating whether to print the names of the excluded variables. Default is `FALSE`.}
}
\value{
A data frame with the specified number of variables starting from the last variable.
}
\description{
Extract a specified number of variables starting from the last variable in a data frame.
}
\examples{
df <- data.frame(
  Snack1 = c("Apple", "Carrot sticks", "Yogurt", "Almonds", "Celery"),
  Snack2 = c("Baby carrots", "Kale chips", "String cheese", "Berries", "Bell peppers"),
  Snack3 = c("Trail mix", "Cottage cheese", "Pineapple", "Walnuts", "Broccoli"),
  Snack4 = c("Popcorn", "Hard-boiled eggs", "Watermelon", "Cashews", "Cauliflower"),
  Snack5 = c("Rice cakes", "Greek yogurt", "Oranges", "Hazelnuts", "Cherry tomatoes")
)

seletail(df, 2, selec = TRUE, exclu = TRUE)
}
