% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/recocheck.R
\name{recocheck}
\alias{recocheck}
\title{recocheck function
Check the links between old variables to be recoded and the new recoded variable}
\usage{
recocheck(df, old_vars, new_var)
}
\description{
recocheck function
Check the links between old variables to be recoded and the new recoded variable
}
