% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sumasf.R
\name{sumasf}
\alias{sumasf}
\title{Insert summary(as.factor())}
\usage{
sumasf()
}
\value{
summary(as.factor()) written with the cursor moved into the brackets
}
\description{
Insert summary(as.factor())
}
\examples{
sumasf()
}
